backend/requirements.txt
fastapi
uvicorn[standard]
websockets
python-dotenv
pyserial
google-adk
